Make camel_imapx_stream_number() return a guint64
Some fields like UIDVALIDITY and MODSEQ may well be larger than can fit in
a 32-bit value. RFC4551 makes this abundantly clear in its examples.

+++ camel/providers/imapx/camel-imapx-stream.h
@@ -127,7 +127,7 @@ gint		camel_imapx_stream_text		(CamelIMAPXStream *is,
 						 CamelException *ex);
 
 /* gets a 'number' */
-guint32		 camel_imapx_stream_number	(CamelIMAPXStream *is,
+guint64		 camel_imapx_stream_number	(CamelIMAPXStream *is,
 						 CamelException *ex);
 
 /* skips the rest of a line, including literals, etc */
